requirements.txt
asgiref==3.7.2
DateTime==5.3
Django==5.0
djangorestframework==3.14.0
itsdangerous==2.1.2
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
zope.interface==6.1
cryptography==41.0.7
PyMySQL~=1.1.0